Consolidated ModalContainer and ModalBody, refactored product info section into ModalInfo component
