--- client/src/components/Modal/ModalBody/package.json
@@ -1,3 +1,0 @@
-{
-  "main": "ModalBody.jsx"
-}
