+++ client/src/components/Modal/ModalContainer/ModalContainer.jsx
@@ -1,7 +1,10 @@
 import React, { Component } from 'react';
 import { createPortal } from 'react-dom';
 import styles from './ModalContainer.css';
-import ModalBody from '../ModalBody';
+import ModalImage from '../ModalImage';
+import ModalInfo from '../ModalInfo';
+import ModalBottomLeft from '../ModalBottomLeft';
+import ModalBottomRight from '../ModalBottomRight';
 import Close from './Close.jsx';
 
 const modalRoot = document.getElementById('modal');
@@ -31,23 +34,31 @@ class Modal extends Component {
           <div className={ styles.container1 }>
             <div className={ styles.container2 }>
               <div className={ styles.dialog }>
-                <ModalBody
-                loved={ this.props.loved }
-                handleLove={ this.props.handleLove }
+                <div className={ styles.box }>
+                  <div className={ styles.flexTop }>
+                    <ModalImage image={ this.props.image } badge={ this.props.badge }/>
+                    <ModalInfo
+                    name={ this.props.name }
+                    category={ this.props.category }
+                    size={ this.props.size }
+                    description={ this.props.description }
+                    sku={ this.props.sku }
+                    />
+                  </div>
+                  <div className={ styles.flexBot }>
+                    <ModalBottomLeft
+                    loved={ this.props.loved }
+                    stars={ this.props.stars }
+                    reviews={ this.props.reviews }
+                    loves={ this.props.loves }
+                    exclusive={ this.props.exclusive }/>
+                    <ModalBottomRight
+                    loved={ this.props.loved }
+                    handleLove={ this.props.handleLove }
+                    price={ this.props.price }/>
+                  </div>
+                </div>
 
-                name={ this.props.name }
-                category={ this.props.category }
-                size={ this.props.size }
-                description={ this.props.description }
-                sku={ this.props.sku }
-                stars={ this.props.stars }
-                reviews={ this.props.reviews }
-                badge={ this.props.badge }
-                loves={ this.props.loves }
-                exclusive={ this.props.exclusive }
-                price={ this.props.price }
-                image={ this.props.image }
-                />
                 <button
                 className={ styles.closeButton }
                 onClick={ e => this.props.onClose(e) }>
