+++ client/src/components/Modal/ModalInfo/ModalInfo.css
@@ -1,19 +1,3 @@
-.box {
-  display: block;
-  position: relative;
-  padding-top: 32px;
-  padding-right: 32px;
-  padding-bottom: 32px;
-  padding-left: 32px;
-}
-.flexTop {
-  display: flex;
-  margin-bottom: 24px;
-  margin-left: -12px;
-  margin-right: -12px;
-  flex-flow: row wrap;
-}
-
 .productInfo {
   padding-left: 12px;
   padding-right: 12px;
@@ -87,11 +71,4 @@
 
 .productLink:hover {
   text-decoration: underline;
-}
-
-.flexBot {
-  display: flex;
-  margin-left: -12px;
-  margin-right: -12px;
-  flex-flow: row wrap;
 }
