+++ client/src/components/Modal/ModalInfo/package.json
@@ -1,0 +1,3 @@
+{
+  "main": "ModalInfo.jsx"
+}
