+++ public/index.html
@@ -4,7 +4,7 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
-  <title>Similar Pages</title>
+  <title>Similar Products and You May Also Like</title>
   <link rel='stylesheet' href='./index.css'/>
 </head>
 <body>
